+++ changelog.md
@@ -4,6 +4,7 @@
 
 * Add StdStream(NoStream) to have standard handles closed. [#13](https://github.com/haskell/process/pull/13)
 * Support for Windows `DETACHED_PROCESS` and `setsid` [#32](https://github.com/haskell/process/issues/32)
+* Support for Windows `CREATE_NEW_CONSOLE` [#38](https://github.com/haskell/process/issues/38)
 
 ## 1.2.3.0 *March 2015*